refactor: refactor file names PE-1904
